+++ src/auxillary/index.js
@@ -1,4 +1,5 @@
 export * from './fade.primitive';
+export * from './head-occlusion.component';
 export * from './motion-input.component';
 export * from './nav-mesh-constrained.component';
 export * from './rotation-input.component';